+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 Language Support:
 
 1.  MarkLogic 7.0 and 8.0 schema-components `KindType` tests.
-1.  Saxon 9.8 `tuple` and `union` types.
+1.  Saxon 9.8 `tuple` and `union` types, and `declare type` prolog statements.
 1.  Recover when `=` is used instead of `:=`.
 1.  Improved multi-statement expression parsing and error reporting for MarkLogic 6.0, and
     W3C Scripting XQuery extensions.